final and completed version

+++ app/src/main/java/projects/shahabgt/com/onlinelibrary/classes/DateParser.java
@@ -58,16 +58,16 @@ public class DateParser {
             gy-=621;
         }
         int gy2 = (gm > 2)?(gy + 1):gy;
-        int days = (365 * gy) + ((int)((gy2 + 3) / 4)) - ((int)((gy2 + 99) / 100)) + ((int)((gy2 + 399) / 400)) - 80 + gd + g_d_m[gm - 1];
-        jy += 33 * ((int)(days / 12053));
+        int days = (365 * gy) + (gy2 + 3) / 4 - (gy2 + 99) / 100 + (gy2 + 399) / 400 - 80 + gd + g_d_m[gm - 1];
+        jy += 33 * days / 12053;
         days %= 12053;
-        jy += 4 * ((int)(days / 1461));
+        jy += 4 * days / 1461;
         days %= 1461;
         if(days > 365){
-            jy+=(int)((days-1)/365);
+            jy+= (days-1)/365;
             days=(days-1)%365;
         }
-        int jm = (days < 186)?1 + (int)(days / 31):7 + (int)((days - 186) / 30);
+        int jm = (days < 186)?1 + days / 31 :7 + (days - 186) / 30;
         int jd = 1 + ((days < 186)?(days % 31):((days - 186) % 30));
